+++ apps/web/src/app.module.ts
@@ -1,7 +1,7 @@
 import { createModule } from '@viness/react'
-import { routerModule } from './routes/routes.module'
-import { dashboardModule } from './features/dashboard/dashboard.module'
+import { AppRouteModule } from './routes/routes.module'
+import { DashboardModule } from './features/dashboard/dashboard.module'
 
 export const appModule = createModule({
-    imports: [routerModule, dashboardModule]
+    imports: [AppRouteModule, DashboardModule]
 })