+++ apps/web/src/features/dashboard/dashboard.module.ts
@@ -1,16 +1,8 @@
-import { createModule } from '@viness/react'
-import { CounterActions } from './store/counter-store'
+import { Module } from '@viness/react'
+import { CounterActions, CounterStore } from './store/counter-store'
 import { Test2Service, TestService } from './services/test.service'
-import { CounterStore, ICounterStore } from './store/counter-store.protocol'
 
-export const dashboardModule = createModule({
-    providers: [
-        {
-            provide: ICounterStore,
-            useClass: CounterStore
-        },
-        CounterActions,
-        TestService,
-        Test2Service
-    ]
+@Module({
+    providers: [CounterStore, CounterActions, TestService, Test2Service]
 })
+export class DashboardModule {}