+++ packages/viness-react/src/app/annotation/injectable.ts
@@ -4,11 +4,15 @@ export const INJECTABLE_ID = '$viness:service_id'
 
 export const tokenToService = new Map<ServiceIdentifier<any>, any>()
 
+export interface Metadata {
+    id: ServiceIdentifier<any>
+}
+
 /**
  *
  */
-export function Injectable(identifier: ServiceIdentifier<any>): ClassDecorator {
+export function Injectable(metadata: ServiceIdentifier<any> | Metadata): ClassDecorator {
     return (target: any) => {
-        Object.defineProperty(target, INJECTABLE_ID, { value: identifier })
+        Object.defineProperty(target, INJECTABLE_ID, { value: metadata })
     }
 }