+++ packages/viness-react/src/router/route-tree.ts
@@ -2,6 +2,7 @@ import { RouteObject, redirect } from 'react-router-dom'
 import { VinessRouteInjectionToken } from './route.protocol'
 import { VinessRoute } from './route'
 import { Injectable } from '../app'
+import { SyncDescriptor } from '@viness/di'
 
 export type RouteToken<P extends string> = VinessRouteInjectionToken<P>
 
@@ -12,27 +13,35 @@ export type RouteNode<P extends string> =
 
 export type RouteTree = RouteNode<any>[]
 
-function toVinessRouteProvider(token: VinessRouteInjectionToken<any>): any {
-    @Injectable(token)
-    class NewVinessRoute extends VinessRoute<any> {}
-    return NewVinessRoute
+function toVinessRouteProvider(token: VinessRouteInjectionToken<any>, path?: string): any {
+    const args = path ? { ...token.metadata, path } : token.metadata
+
+    return { provide: token, useClass: new SyncDescriptor(VinessRoute, [args]) }
 }
 
-export function convertToVinessRouteProviders(tree: RouteTree) {
+export function convertToVinessRouteProviders(tree: RouteTree, parentPath: string[] = []) {
     const routes: any[] = []
     tree.forEach((node) => {
         if (typeof node === 'function') {
             node as RouteToken<any>
 
-            const route = toVinessRouteProvider(node)
-            routes.push(route)
+            const metadata = node.metadata
+            const path = metadata.path as string
+            if (path.startsWith('/')) {
+                const route = toVinessRouteProvider(node)
+                routes.push(route)
+            } else {
+                const newPath = [...parentPath, path].join('/')
+                const route = toVinessRouteProvider(node, newPath)
+                routes.push(route)
+            }
         } else if (Array.isArray(node)) {
-            const [_, children] = node
-            const newRoutes = convertToReactRoutes(children)
+            const [parent, children] = node
+            const newRoutes = convertToVinessRouteProviders(children, [...parentPath, parent.metadata.path])
             routes.push(...newRoutes)
         } else {
             const { children } = node
-            const newRoutes = convertToReactRoutes(children)
+            const newRoutes = convertToVinessRouteProviders(children, [...parentPath, node.route.metadata.path])
             routes.push(...newRoutes)
         }
     })
@@ -51,7 +60,9 @@ export function convertToReactRoutes(tree: RouteTree): RouteObject[] {
         if (typeof node === 'function') {
             node as RouteToken<any>
 
-            return node.metadata
+            const routeObject = node.metadata
+
+            return routeObject
         } else if (Array.isArray(node)) {
             const [route, children] = node
             const newChildren = convertToReactRoutes(children)