+++ packages/viness-react/src/router/route.factory.ts
@@ -1,4 +1,5 @@
 import { TokenType, createToken } from '../token'
+import { generateId } from '../utils'
 import { IVinessRoute, VinessRouteInjectionToken, VinessRouteMetadata } from './route.protocol'
 
 /**
@@ -9,7 +10,9 @@ import { IVinessRoute, VinessRouteInjectionToken, VinessRouteMetadata } from './
  * @returns
  */
 export function createRoute<Path extends string>(path: Path, metadata?: Omit<VinessRouteMetadata<Path>, 'path'>) {
-    const token = createToken<IVinessRoute<Path>>(path || 'route') as VinessRouteInjectionToken<Path>
+    // generate an id to make sure the service id unique
+    const id = generateId()
+    const token = createToken<IVinessRoute<Path>>(`${path || 'route'}_${id}`) as VinessRouteInjectionToken<Path>
     token.metadata = { ...metadata, path }
     token.tokenType = TokenType.Route
     return token