+++ packages/viness-react/src/router/router.module.ts
@@ -4,6 +4,7 @@ import { createModule } from '../app/module/module'
 import { RouterParams, VinessRouter } from './router'
 import { IVinessRouter, RouterConfig } from './router.protocol'
 import { convertToReactRoutes, convertToVinessRouteProviders } from './route-tree'
+import { doLog } from '../utils'
 
 export function createRouterModule(config: RouterConfig) {
     const { type, basename, routeTree } = config
@@ -12,6 +13,8 @@ export function createRouterModule(config: RouterConfig) {
     const routeProviders = convertToVinessRouteProviders(routeTree)
 
     const routerParams: RouterParams = { type, routes, basename }
+
+    doLog(routes)
 
     const descriptor = new SyncDescriptor(VinessRouter, [routerParams], true)
     ContaienrUtil.register(IVinessRouter, descriptor)